+++ core/thread.py
@@ -24,10 +24,10 @@ class Thread:
         self.channel = None
         self.ready_event = asyncio.Event()
         self.close_task = None
-        self.close_after = 0 # seconds
 
     def __repr__(self):
-        return f'Thread(recipient="{self.recipient}", channel={self.channel.id})'
+        return f'Thread(recipient="{self.recipient}", ' \
+               f'channel={self.channel.id})'
 
     def wait_until_ready(self):
         """Blocks execution until the thread is fully set up."""
@@ -42,28 +42,54 @@ class Thread:
         if flag is True:
             self.ready_event.set()
     
-    async def _close_after(self, after, **kwargs):
-        await asyncio.sleep(after)
-        kwargs['scheduled'] = True
-        await self.close(**kwargs)
+    def _close_after(self, closer, silent, delete_channel, message):
+        return self.bot.loop.create_task(
+            self._close(closer, silent, delete_channel, message, True))
 
-    async def close(self, *, closer, after=0, silent=False, delete_channel=True, message=None, scheduled=False):
-        '''Close a thread now or after a set time in seconds'''
-        if self.close_task is not None and not self.close_task.cancelled():
-            if not scheduled or after > 0:
-                self.close_task.cancel()
+    async def close(self, *, closer, after=0, silent=False,
+                    delete_channel=True, message=None):
+        """Close a thread now or after a set time in seconds"""
+
+        # restarts the after timer
+        await self.cancel_closure()
 
         if after > 0:
-            self.close_task = asyncio.create_task(self._close_after(after, closer=closer, silent=silent, message=message))
-            return 
+            # TODO: Add somewhere to clean up broken closures
+            #  (when channel is already deleted)
+            await self.bot.config.update()
+            now = datetime.datetime.utcnow()
+            items = {
+                # 'initiation_time': now.isoformat(),
+                'time': (now + datetime.timedelta(seconds=after)).isoformat(),
+                'closer_id': closer.id,
+                'silent': silent,
+                'delete_channel': delete_channel,
+                'message': message
+            }
+            self.bot.config.closures[str(self.id)] = items
+            await self.bot.config.update()
 
+            self.close_task = self.bot.loop.call_later(
+                after, self._close_after, closer,
+                silent, delete_channel, message)
+            return
+
+        return await self._close(closer, silent, delete_channel, message)
+
+    async def _close(self, closer, silent=False, delete_channel=True,
+                     message=None, scheduled=False):
         del self.manager.cache[self.id]
+
+        await self.cancel_closure()
+
         if str(self.id) in self.bot.config.subscriptions:
             del self.bot.config.subscriptions[str(self.id)]
 
         # Logging
         log_data = await self.bot.modmail_api.post_log(self.channel.id, {
-            'open': False, 'closed_at': str(datetime.datetime.utcnow()), 'closer': {
+            'open': False,
+            'closed_at': str(datetime.datetime.utcnow()),
+            'closer': {
                 'id': str(closer.id),
                 'name': closer.name,
                 'discriminator': closer.discriminator,
@@ -73,12 +99,14 @@ class Thread:
         })
 
         if isinstance(log_data, str):
-            print(log_data) # errored somehow on server
+            print(log_data)  # errored somehow on server
+            return
 
         if self.bot.selfhosted:
             log_url = f'{self.bot.config.log_url}/logs/{log_data["key"]}'
         else:
-            log_url = f"https://logs.modmail.tk/{log_data['user_id']}/{log_data['key']}" 
+            log_url = f"https://logs.modmail.tk/" \
+                      f"{log_data['user_id']}/{log_data['key']}"
 
         user = self.recipient.mention if self.recipient else f'`{self.id}`'
 
@@ -104,9 +132,9 @@ class Thread:
 
         # Thread closed message 
 
-        em = discord.Embed(title='Thread Closed')
-        em.description = message or f'{closer.mention} has closed this modmail thread.'
-        em.color = discord.Color.red()
+        em = discord.Embed(title='Thread Closed', color=discord.Color.red())
+        em.description = message or \
+            f'{closer.mention} has closed this modmail thread.'
 
         if not silent and self.recipient is not None:
             tasks.append(self.recipient.send(embed=em))
@@ -116,7 +144,17 @@ class Thread:
         
         await asyncio.gather(*tasks)
 
-    async def _edit_thread_message(self, channel, message_id, message):
+    async def cancel_closure(self):
+        if self.close_task is not None:
+            self.close_task.cancel()
+            self.close_task = None
+
+        to_update = self.bot.config.closures.pop(str(self.id), None)
+        if to_update is not None:
+            await self.bot.config.update()
+
+    @staticmethod
+    async def _edit_thread_message(channel, message_id, message):
         async for msg in channel.history():
             if not msg.embeds:
                 continue
@@ -124,7 +162,8 @@ class Thread:
             if embed and embed.author and embed.author.url:
                 if str(message_id) == str(embed.author.url).split('/')[-1]:
                     if ' - (Edited)' not in embed.footer.text:
-                        embed.set_footer(text=embed.footer.text + ' - (Edited)')
+                        embed.set_footer(
+                            text=embed.footer.text + ' - (Edited)')
                     embed.description = message
                     await msg.edit(embed=embed)
                     break
@@ -139,31 +178,47 @@ class Thread:
         if not message.content and not message.attachments:
             raise commands.UserInputError
         if all(not g.get_member(self.id) for g in self.bot.guilds):
-            return await message.channel.send(embed=discord.Embed(color=discord.Color.red(), description='This user shares no servers with me and is thus unreachable.'))
+            return await message.channel.send(
+                embed=discord.Embed(
+                    color=discord.Color.red(),
+                    description='This user shares no servers with '
+                                'me and is thus unreachable.'))
 
         tasks = [
-            self.send(message, self.channel, from_mod=True),  # in thread channel
-            self.send(message, self.recipient, from_mod=True)  # to user
+            # in thread channel
+            self.send(message, self.channel, from_mod=True),
+            # to user
+            self.send(message, self.recipient, from_mod=True)
             ]
 
-        if self.close_task is not None and not self.close_task.cancelled():
-            self.close_task.cancel() # cancel closing if a thread message is sent.
-            tasks.append(self.channel.send(embed=discord.Embed(color=discord.Color.red(), description='Scheduled close has been cancelled.')))
+        if self.close_task is not None:
+            # cancel closing if a thread message is sent.
+            await self.cancel_closure()
+            tasks.append(self.channel.send(
+                embed=discord.Embed(color=discord.Color.red(),
+                                    description='Scheduled close has '
+                                                'been cancelled.')))
 
         await asyncio.gather(*tasks)
 
-    async def send(self, message, destination=None, from_mod=False, delete_message=True):
-        if self.close_task is not None and not self.close_task.cancelled():
-            self.close_task.cancel() # cancel closing if a thread message is sent.
-            await self.channel.send(embed=discord.Embed(color=discord.Color.red(), description='Scheduled close has been cancelled.'))
+    async def send(self, message, destination=None, from_mod=False):
+        if self.close_task is not None:
+            # cancel closing if a thread message is sent.
+            await self.cancel_closure()
+            await self.channel.send(embed=discord.Embed(
+                color=discord.Color.red(),
+                description='Scheduled close has been cancelled.'))
+
         if not self.ready:
             await self.wait_until_ready()
 
         destination = destination or self.channel
         if from_mod and not isinstance(destination, discord.User):
-            asyncio.create_task(self.bot.modmail_api.append_log(message))
+            self.bot.loop.create_task(
+                self.bot.modmail_api.append_log(message))
         elif not from_mod:
-            asyncio.create_task(self.bot.modmail_api.append_log(message, destination.id))
+            self.bot.loop.create_task(
+                self.bot.modmail_api.append_log(message, destination.id))
 
         author = message.author
 
@@ -172,10 +227,16 @@ class Thread:
             timestamp=message.created_at
         )
 
-        em.set_author(name=str(author), icon_url=author.avatar_url, url=message.jump_url)  # store message id in hidden url
+        # store message id in hidden url
+        em.set_author(name=str(author),
+                      icon_url=author.avatar_url,
+                      url=message.jump_url)
 
         image_types = ['.png', '.jpg', '.gif', '.jpeg', '.webp']
-        is_image_url = lambda u, _: any(urlparse(u.lower()).path.endswith(x) for x in image_types)
+
+        def is_image_url(u, _):
+            return any(urlparse(u.lower()).path.endswith(x)
+                       for x in image_types)
 
         delete_message = not bool(message.attachments)
 
@@ -184,7 +245,8 @@ class Thread:
         images = [x for x in attachments if is_image_url(*x)]
         attachments = [x for x in attachments if not is_image_url(*x)]
 
-        image_links = [(link, None) for link in re.findall(r'(https?://[^\s]+)', message.content)]
+        image_links = [(link, None) for link in
+                       re.findall(r'(https?://[^\s]+)', message.content)]
         image_links = [x for x in image_links if is_image_url(*x)]
         images.extend(image_links)
 
@@ -195,19 +257,24 @@ class Thread:
         additional_count = 1
 
         for att in images:
-            if is_image_url(*att) and not embedded_image and att[1] if prioritize_uploads else True:
+            if is_image_url(*att) and not embedded_image and att[1] \
+                    if prioritize_uploads else True:
                 em.set_image(url=att[0])
                 embedded_image = True
             elif att[1] is not None:
                 link = f'[{att[1]}]({att[0]})'
-                em.add_field(name=f'Additional Image upload ({additional_count})', value=link, inline=False)
+                em.add_field(
+                    name=f'Additional Image upload ({additional_count})',
+                    value=link,
+                    inline=False
+                )
                 additional_count += 1
         
         file_upload_count = 1
 
-
         for att in attachments:
-            em.add_field(name=f'File upload ({file_upload_count})', value=f'[{att[1]}]({att[0]})')
+            em.add_field(name=f'File upload ({file_upload_count})',
+                         value=f'[{att[1]}]({att[0]})')
             file_upload_count += 1
 
         if from_mod:
@@ -242,7 +309,7 @@ class Thread:
         if key in config['notification_squad']:
             mentions.extend(config['notification_squad'][key])
             del config['notification_squad'][key]
-            asyncio.create_task(config.update())
+            self.bot.loop.create_task(config.update())
         
         return ' '.join(mentions)
 
@@ -256,7 +323,8 @@ class ThreadManager:
 
     async def populate_cache(self):
         for channel in self.bot.modmail_guild.text_channels:
-            if not self.bot.using_multiple_server_setup and channel.category != self.bot.main_category:
+            if not self.bot.using_multiple_server_setup and \
+                    channel.category != self.bot.main_category:
                 continue
             await self.find(channel=channel)
 
@@ -273,6 +341,8 @@ class ThreadManager:
         """Finds a thread from cache or from discord channel topics."""
         if recipient is None and channel is not None:
             return await self._find_from_channel(channel)
+
+        thread = None
         try:
             thread = self.cache[recipient.id]
         except KeyError:
@@ -280,10 +350,9 @@ class ThreadManager:
                 self.bot.modmail_guild.text_channels,
                 topic=f'User ID: {recipient.id}'
             )
-            if not channel:
-                thread = None
-            else:
+            if channel:
                 self.cache[recipient.id] = thread = Thread(self, recipient)
+                # TODO: Fix this:
                 thread.channel = channel
                 thread.ready = True
         finally:
@@ -293,19 +362,22 @@ class ThreadManager:
         """
         Tries to find a thread from a channel channel topic,
         if channel topic doesnt exist for some reason, falls back to
-        searching channel history for genesis embed and extracts user_id fron that.
+        searching channel history for genesis embed and
+        extracts user_id from that.
         """
         user_id = None
 
         if channel.topic and 'User ID: ' in channel.topic:
             user_id = int(re.findall(r'\d+', channel.topic)[0])
 
-        # BUG: When discord fails to create channel topic. search through message history
+        # BUG: When discord fails to create channel topic.
+        # search through message history
         elif channel.topic is None:
             async for message in channel.history(limit=50):
                 if message.embeds:
                     em = message.embeds[0]
-                    matches = re.findall(r'User ID: (\d+)', str(em.footer.text))
+                    # TODO: use re.search instead
+                    matches = re.findall(r'User ID: (\d+)', em.footer.text)
                     if matches:
                         user_id = int(matches[0])
                         break
@@ -328,12 +400,15 @@ class ThreadManager:
 
         em = discord.Embed(
             title='Thread created!',
-            description=self.bot.config.get('thread_creation_response', 'The moderation team will get back to you as soon as possible!'),
+            description=self.bot.config.get(
+                'thread_creation_response',
+                'The moderation team will get back to you as soon as possible!'
+            ),
             color=discord.Color.green()
         )
 
         if creator is None:
-            asyncio.create_task(recipient.send(embed=em))
+            self.bot.loop.create_task(recipient.send(embed=em))
 
         self.cache[recipient.id] = thread = Thread(self, recipient)
 
@@ -345,16 +420,20 @@ class ThreadManager:
         thread.channel = channel
 
         log_url, log_data = await asyncio.gather(
-            self.bot.modmail_api.get_log_url(recipient, channel, creator or recipient),
+            self.bot.modmail_api.get_log_url(recipient, channel,
+                                             creator or recipient),
             self.bot.modmail_api.get_user_logs(recipient.id)
             # self.get_dominant_color(recipient.avatar_url)
         )
 
         log_count = sum(1 for log in log_data if not log['open'])
-        info_embed = self._format_info_embed(recipient, creator, log_url, log_count, discord.Color.green())
+        info_embed = self._format_info_embed(recipient, creator,
+                                             log_url, log_count,
+                                             discord.Color.green())
 
         topic = f'User ID: {recipient.id}'
-        mention = self.bot.config.get('mention', '@here') if not creator else None
+        mention = self.bot.config.get('mention', '@here') \
+            if not creator else None
 
         _, msg = await asyncio.gather(
             channel.edit(topic=topic),
@@ -368,7 +447,8 @@ class ThreadManager:
         return thread
 
     async def find_or_create(self, recipient):
-        return await self.find(recipient=recipient) or await self.create(recipient)
+        return await self.find(recipient=recipient) or \
+               await self.create(recipient)
 
     @staticmethod
     def valid_image_url(url):
@@ -397,7 +477,7 @@ class ThreadManager:
             async with self.bot.session.get(url) as resp:
                 image = await resp.read()
                 color = await self._do_get_dc(image, quality)
-        except Exception:
+        except:
             traceback.print_exc()
             return discord.Color.blurple()
         else:
@@ -410,7 +490,8 @@ class ThreadManager:
         new_name = ''.join(l for l in name if l in allowed) or 'null'
         new_name += f'-{author.discriminator}'
 
-        while new_name in [c.name for c in self.bot.modmail_guild.text_channels]:
+        while new_name in [c.name for c in
+                           self.bot.modmail_guild.text_channels]:
             new_name += '-x'  # two channels with same name
 
         return new_name
@@ -421,21 +502,26 @@ class ThreadManager:
         member = self.bot.guild.get_member(user.id)
         avi = user.avatar_url
         time = datetime.datetime.utcnow()
-        desc = f'{creator.mention} has created a thread with {user.mention}' if creator else f'{user.mention} has started a thread'
+        desc = f'{creator.mention} has created a thread with {user.mention}' \
+            if creator else f'{user.mention} has started a thread'
         key = log_url.split('/')[-1]
         desc = f'{desc} [`{key}`]({log_url})'
 
+        role_names = ''
         if member:
             seperate_server = self.bot.guild != self.bot.modmail_guild
             roles = sorted(member.roles, key=lambda c: c.position)
             if seperate_server:
-                rolenames = ', '.join(r.name for r in roles if r.name != "@everyone")
+                role_names = ', '.join(r.name for r in roles
+                                       if r.name != "@everyone")
             else:
-                rolenames = ' '.join(r.mention for r in roles if r.name != "@everyone")
+                role_names = ' '.join(r.mention for r in roles
+                                      if r.name != "@everyone")
 
         em = discord.Embed(colour=dc, description=desc, timestamp=time)
 
-        days = lambda d: (' day ago.' if d == '1' else ' days ago.')
+        def days(d):
+            return ' day ago.' if d == '1' else ' days ago.'
 
         created = str((time - user.created_at).days)
         # em.add_field(name='Mention', value=user.mention)
@@ -452,9 +538,11 @@ class ThreadManager:
             em.add_field(name='Joined', value=joined + days(joined))
             if member.nick:
                 em.add_field(name='Nickname', value=member.nick, inline=True)
-            if rolenames:
-                em.add_field(name='Roles', value=rolenames, inline=False)
+            if role_names:
+                em.add_field(name='Roles', value=role_names, inline=False)
         else:
-            em.set_footer(text=footer + ' | Note: this member is not part of this server.')
+            em.set_footer(
+                text=f'{footer} | Note: this member'
+                     f' is not part of this server.')
 
         return em